Remove obsolete Qt desktop code
