+++ .gitignore
@@ -1,17 +1,17 @@
 /.venv/
 /.venv312/
 /.idea/
-/.__pycache__/
+__pycache__/
 *.pyc
 /sessions.db
 /recordings/
+/reports/
+/media/
 /scripts/
 /script_index.json
 /settings.json
-/SpeechPractice.spec
 /build/
 /dist/
-/run_app.bat
 /.claude
 Agents..md
 CLAUDE.md
