+++ requirements.txt
@@ -1,8 +1,5 @@
-PyQt5>=5.15
 Django>=5.0,<6.0
-sounddevice>=0.4
 numpy>=1.24
-pyqtgraph>=0.13
 openai-whisper>=20231117
 # Cloud generation providers:
 openai>=1.66
@@ -14,10 +11,7 @@ cryptography>=42.0
 requests>=2.31
 soundfile>=0.12
 pydub>=0.25.1
-# Optional: packaging helper
-pyinstaller>=6.0
 # if running python 3.12 or lower, and have an Nvidia GPU you can install CUDA enabled torch to run whisper on:
 # (GPU): pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu12
 
-# Build into executable: pyinstaller -y --windowed --name SpeechPractice --add-data "scripts;scripts" --add-data "script_index.json;." SpeechPractice.py
 # Ensure ffmpeg is on PATH for Whisper
