--- transcription_service.py
@@ -1,1009 +1,0 @@
-from __future__ import annotations
-
-import json
-import os
-import time
-import gc
-from dataclasses import dataclass
-from datetime import datetime
-from typing import Optional, Any, Dict, Tuple, List
-
-import whisper
-from PyQt5 import QtCore
-from PyQt5.QtWidgets import QListWidgetItem
-from jiwer import cer as jiwer_cer
-
-import db
-from transcribe_worker import TranscribeWorker, FreeTranscribeWorker
-from transcript_utils import (
-    build_transcript_from_segments,
-    highlight_transcript_at_time,
-)
-from settings_ui import whisper_options
-from alignment_utils import (
-    compute_error_spans_for_display,
-    extract_mistake_pairs_for_display,
-)
-from error_analytics import extract_error_events
-
-
-@dataclass
-class TranscriptionRunContext:
-    mode: str
-    session_id: Optional[int]
-    script_name: str
-    script_text: str
-    audio_path: Optional[str] = None
-    worker: Optional[Any] = None
-    received_segments: bool = False
-
-
-class TranscriptionService(QtCore.QObject):
-    """
-    Centralized service for handling transcription and scoring.
-    Adds extended metrics (CER, fluency, confidence) and error highlights.
-    """
-
-    def __init__(self, window):
-        super().__init__()
-        self.window = window
-        self.model: Optional[Any] = None
-        self._received_segments_this_run: bool = False
-        self._active_score_context: Optional[TranscriptionRunContext] = None
-        self._active_free_context: Optional[TranscriptionRunContext] = None
-
-    # ------------------------ model/options ------------------------
-
-    def _resolve_model_device(self) -> str:
-        """
-        Resolve Whisper model device from settings:
-        - gpu  -> cuda (fallback cpu if unavailable)
-        - cpu  -> cpu
-        - auto -> cuda if available else cpu
-        """
-        pref = (
-            getattr(self.window, "settings", {}).get("device", "auto")
-            if hasattr(self.window, "settings")
-            and isinstance(self.window.settings, dict)
-            else "auto"
-        )
-        if pref == "cpu":
-            return "cpu"
-        try:
-            import torch
-
-            has_cuda = bool(torch.cuda.is_available())
-        except Exception:
-            has_cuda = False
-        if pref == "gpu":
-            return "cuda" if has_cuda else "cpu"
-        return "cuda" if has_cuda else "cpu"
-
-    def ensure_model(self) -> None:
-        if self.model is None:
-            self._release_finished_workers()
-            gc.collect()
-            self._clear_cuda_cache()
-            model_name = (
-                getattr(self.window, "settings", {}).get("model_name")
-                if hasattr(self.window, "settings")
-                and isinstance(self.window.settings, dict)
-                else None
-            ) or os.getenv("WHISPER_MODEL", "base.en")
-            device = self._resolve_model_device()
-            try:
-                self.model = whisper.load_model(model_name, device=device)
-            except RuntimeError as e:
-                if "out of memory" in str(e).lower():
-                    self._clear_cuda_cache()
-                    gc.collect()
-                    raise RuntimeError(
-                        "CUDA out of memory while loading Whisper model. "
-                        "Try a smaller model or wait for current transcriptions "
-                        "to finish before switching models."
-                    ) from e
-                raise
-            try:
-                self.model.eval()
-            except Exception:
-                pass
-
-    def get_whisper_options(self, free_speak: bool = False) -> dict:
-        return whisper_options(self.window.settings, free_speak=free_speak)
-
-    def _get_worker_timing(
-        self, free_speak: bool = False, worker: Optional[Any] = None
-    ) -> Dict[str, float]:
-        if worker is None:
-            worker_attr = "free_worker" if free_speak else "worker"
-            worker = getattr(self.window, worker_attr, None)
-        timing = getattr(worker, "last_timing", None)
-        return dict(timing) if isinstance(timing, dict) else {}
-
-    @staticmethod
-    def _format_timing_text(timing: Dict[str, float], ui_post_s: float) -> str:
-        asr_s = float(timing.get("asr_s", 0.0))
-        worker_post_s = float(timing.get("worker_post_s", 0.0))
-        post_s = max(0.0, worker_post_s) + max(0.0, ui_post_s)
-        total_s = max(0.0, asr_s) + post_s
-        return f"Timing: ASR {asr_s:.2f}s | Post {post_s:.2f}s | Total {total_s:.2f}s"
-
-    def _set_timing_text(self, text: str) -> None:
-        try:
-            if (
-                hasattr(self.window, "timing_label")
-                and self.window.timing_label is not None
-            ):
-                self.window.timing_label.setText(str(text))
-        except Exception:
-            pass
-
-    def _clear_cuda_cache(self) -> None:
-        try:
-            import torch
-
-            if torch.cuda.is_available():
-                torch.cuda.empty_cache()
-                try:
-                    torch.cuda.ipc_collect()
-                except Exception:
-                    pass
-        except Exception:
-            pass
-
-    def _release_finished_workers(self) -> None:
-        for attr in ("worker", "free_worker"):
-            try:
-                w = getattr(self.window, attr, None)
-                if w is not None and not w.isRunning():
-                    setattr(self.window, attr, None)
-            except Exception:
-                pass
-
-    def is_busy(self) -> bool:
-        for attr in ("worker", "free_worker"):
-            try:
-                w = getattr(self.window, attr, None)
-                if w is not None and w.isRunning():
-                    return True
-            except Exception:
-                pass
-        return False
-
-    def can_start_transcription(self) -> bool:
-        if self.is_busy():
-            return False
-        if bool(getattr(self.window, "free_speak_mode", False)):
-            return (
-                getattr(self.window, "audio_data", None) is not None
-                or getattr(self.window, "current_audio_path", None) is not None
-            )
-        return getattr(self.window, "current_audio_path", None) is not None
-
-    def _is_context_visible(self, context: Optional[TranscriptionRunContext]) -> bool:
-        if context is None:
-            return True
-        current_session_id = getattr(self.window, "current_session_id", None)
-        if context.session_id is not None:
-            return current_session_id == context.session_id
-        if context.mode == "free":
-            return (
-                current_session_id is None
-                and bool(getattr(self.window, "free_speak_mode", False))
-            )
-        return (
-            current_session_id is None
-            and getattr(self.window, "current_audio_path", None) == context.audio_path
-            and getattr(self.window, "current_script_text", "") == context.script_text
-        )
-
-    def _add_history_item(self, sess) -> None:
-        formatted_timestamp = datetime.strptime(
-            sess.timestamp, "%Y-%m-%dT%H:%M:%S"
-        ).strftime("%d %b %Y %H:%M")
-        label = f"{formatted_timestamp} - {sess.script_name}"
-        it = QListWidgetItem(label)
-        it.setData(QtCore.Qt.UserRole, sess.id)
-        try:
-            it.setIcon(self.window._session_type_icon(sess.script_name))
-        except Exception:
-            pass
-        if self.window.history_list.count() == 1 and not isinstance(
-            self.window.history_list.item(0).data(QtCore.Qt.UserRole),
-            int,
-        ):
-            self.window.history_list.takeItem(0)
-        self.window.history_list.addItem(it)
-
-    def _set_background_status(self, text: str | None) -> None:
-        try:
-            if hasattr(self.window, "set_background_status"):
-                self.window.set_background_status(text)
-        except Exception:
-            pass
-
-    def _sync_export_button(self) -> None:
-        try:
-            if hasattr(self.window, "_sync_export_button"):
-                self.window._sync_export_button()
-        except Exception:
-            pass
-
-    def _mark_run_started(self, context: TranscriptionRunContext) -> None:
-        try:
-            if context.mode == "score":
-                self.window.mark_session_scoring(context.session_id, True)
-                self._set_background_status(
-                    f"Scoring in background: {context.script_name}"
-                )
-            else:
-                self._set_background_status("Transcribing Free Speak in background")
-        except Exception:
-            pass
-
-    def _mark_score_run_finished(
-        self,
-        context: Optional[TranscriptionRunContext],
-        sess,
-        failed: bool = False,
-    ) -> None:
-        if context is not None:
-            try:
-                self.window.mark_session_scoring(context.session_id, False)
-            except Exception:
-                pass
-        if sess is not None:
-            try:
-                self.window.refresh_history_session(sess)
-            except Exception:
-                pass
-            try:
-                when = self.window._session_timestamp_label(sess.timestamp)
-            except Exception:
-                when = str(getattr(sess, "timestamp", "") or "")
-            if failed:
-                self._set_background_status(f"Scoring failed for session from {when}")
-            else:
-                self._set_background_status(f"Finished scoring session from {when}")
-        elif failed:
-            self._set_background_status("Scoring failed")
-        else:
-            self._set_background_status("Finished scoring")
-
-    def _mark_free_run_finished(self, failed: bool = False) -> None:
-        if failed:
-            self._set_background_status("Free Speak transcription failed")
-        else:
-            self._set_background_status("Free Speak transcript ready")
-
-    def _finish_worker(self) -> None:
-        self._release_finished_workers()
-        try:
-            self.window.btn_score.setEnabled(self.can_start_transcription())
-        except Exception:
-            pass
-
-    def unload_model(self) -> None:
-        self.model = None
-        self._release_finished_workers()
-        gc.collect()
-        self._clear_cuda_cache()
-
-    # ------------------------ metrics helpers ----------------------
-
-    @staticmethod
-    def _clean_text_for_metrics(text: str) -> str:
-        # Keep consistent with worker normalization
-        return TranscribeWorker.clean_text(text)
-
-    def _compute_cer(self, ref_text: str, hyp_text: str) -> float:
-        ref = self._clean_text_for_metrics(ref_text).replace(" ", "")
-        hyp = self._clean_text_for_metrics(hyp_text).replace(" ", "")
-        try:
-            return float(jiwer_cer(ref, hyp))
-        except Exception:
-            return 0.0
-
-    @staticmethod
-    def _norm_conf_from_logprob(lp: Optional[float]) -> Optional[float]:
-        if lp is None:
-            return None
-        # Whisper avg_logprob is typically in [-1, 0]; map to [0, 1].
-        conf = (float(lp) + 1.0) / 1.0
-        return max(0.0, min(1.0, conf))
-
-    def _augment_segments_and_fluency(
-        self, segments: List[dict], hyp_text: str
-    ) -> Tuple[List[dict], float, float, float, Optional[float]]:
-        """
-        Add duration, pause_before, conf to segments.
-        Return (segments_aug, articulation_rate_wpm, pause_ratio,
-                filled_pauses_count, avg_conf)
-        """
-        segs: List[dict] = []
-        prev_end: Optional[float] = None
-        speech_time = 0.0
-        pause_time = 0.0
-        conf_vals: List[float] = []
-
-        for seg in segments or []:
-            start = float(seg.get("start", 0.0))
-            end = float(seg.get("end", start))
-            dur = max(0.0, end - start)
-            pb = 0.0
-            if prev_end is not None:
-                pb = max(0.0, start - prev_end)
-                pause_time += pb
-            prev_end = end
-            speech_time += dur
-
-            lp = seg.get("avg_logprob", None)
-            conf = self._norm_conf_from_logprob(lp)
-            if conf is not None:
-                conf_vals.append(conf)
-
-            seg_copy = dict(seg)
-            seg_copy["duration"] = float(dur)
-            seg_copy["pause_before"] = float(pb)
-            if conf is not None:
-                seg_copy["conf"] = float(conf)
-            segs.append(seg_copy)
-
-        total_time = 0.0
-        if segments:
-            total_time = float(
-                max(
-                    0.0,
-                    float(segments[-1].get("end", 0.0))
-                    - float(segments[0].get("start", 0.0)),
-                )
-            )
-
-        # Words per minute (exclude pauses): words / (speech_time / 60)
-        hyp_clean = self._clean_text_for_metrics(hyp_text)
-        n_words = len(hyp_clean.split())
-        artic_rate = 0.0
-        if speech_time > 1e-6:
-            artic_rate = float(n_words) * (60.0 / speech_time)
-
-        pause_ratio = 0.0
-        denom = total_time if total_time > 1e-6 else (speech_time + pause_time)
-        if denom > 1e-6:
-            pause_ratio = pause_time / denom
-
-        # Filled pauses in the hypothesis
-        filled_set = {"um", "uh", "erm", "er", "hmm"}
-        filled_count = float(sum(1 for t in hyp_clean.split() if t in filled_set))
-
-        avg_conf = float(sum(conf_vals) / len(conf_vals)) if conf_vals else None
-
-        return segs, artic_rate, pause_ratio, filled_count, avg_conf
-
-    # ------------------------ error spans --------------------------
-
-    def compute_error_spans(
-        self, script_display_text: str, transcript_display_text: str
-    ) -> Tuple[List[Tuple[int, int, str]], List[Tuple[int, int, str]]]:
-        """
-        Compute script/transcript highlight spans based on word/char alignment.
-        """
-        try:
-            return compute_error_spans_for_display(
-                script_display_text, transcript_display_text
-            )
-        except Exception:
-            return [], []
-
-    # ------------------------ workflows ----------------------------
-
-    def transcribe_and_score(self) -> None:
-        """
-        Script-based scoring. Uses ASR, computes WER/CER/Clarity and
-        extended metrics when timestamps are available.
-        """
-        if self.is_busy():
-            self.window.metrics_label.setText("Transcription already running.")
-            return
-
-        if self.window.current_audio_path is None:
-            self.window.metrics_label.setText("No audio to score.")
-            self._set_timing_text("Timing: -")
-            return
-
-        if self.window.free_speak_mode:
-            self.transcribe_free()
-            return
-
-        self.window.btn_score.setEnabled(False)
-        self.window.metrics_label.setText("Scoring... please wait")
-        self._set_timing_text("Timing: running...")
-        try:
-            self.ensure_model()
-        except Exception as e:
-            self.window.metrics_label.setText(f"Scoring failed: {e}")
-            self._set_timing_text("Timing: failed")
-            self.window.btn_score.setEnabled(True)
-            return
-
-        context = TranscriptionRunContext(
-            mode="score",
-            session_id=getattr(self.window, "current_session_id", None),
-            script_name=getattr(self.window, "current_script_name", ""),
-            script_text=getattr(self.window, "current_script_text", ""),
-            audio_path=getattr(self.window, "current_audio_path", None),
-        )
-        self._active_score_context = context
-        self._mark_run_started(context)
-
-        self.window.worker = TranscribeWorker(
-            self.model,
-            context.script_text,
-            context.audio_path,
-            self.window,
-            options=self.get_whisper_options(),
-        )
-        context.worker = self.window.worker
-        self._received_segments_this_run = False
-
-        # Connect before start to avoid race
-        try:
-            self.window.worker.partial.connect(
-                lambda hyp, ctx=context: self.on_transcription_partial(hyp, ctx)
-            )
-            self.window.worker.completed_with_segments.connect(
-                lambda hyp, err, clar, score, segments, ctx=context: (
-                    self.on_transcription_done_with_segments(
-                        hyp, err, clar, score, segments, ctx
-                    )
-                )
-            )
-        except Exception:
-            pass
-        self.window.worker.completed.connect(
-            lambda hyp, err, clar, score, ctx=context: self.on_transcription_done(
-                hyp, err, clar, score, ctx
-            )
-        )
-        self.window.worker.failed.connect(
-            lambda msg, ctx=context: self.on_worker_failed(msg, ctx)
-        )
-        self.window.worker.finished.connect(self._finish_worker)
-        self.window.worker.start()
-
-    def transcribe_free(self) -> None:
-        """
-        Free Speak: no WER/CER; still compute fluency/confidence
-        when timestamps are available.
-        """
-        if self.is_busy():
-            self.window.metrics_label.setText("Transcription already running.")
-            return
-
-        if self.window.audio_data is None and self.window.current_audio_path is None:
-            self.window.metrics_label.setText("Nothing to transcribe.")
-            self._set_timing_text("Timing: -")
-            return
-
-        self.window.btn_score.setEnabled(False)
-        self.window.metrics_label.setText("Transcribing... (Free Speak)")
-        self._set_timing_text("Timing: running...")
-
-        audio_input = (
-            self.window.audio_data
-            if self.window.audio_data is not None
-            else self.window.current_audio_path
-        )
-
-        try:
-            self.ensure_model()
-        except Exception as e:
-            self.window.metrics_label.setText(f"Transcription failed: {e}")
-            self._set_timing_text("Timing: failed")
-            self.window.btn_score.setEnabled(True)
-            try:
-                self.window.btn_save.setEnabled(False)
-            except Exception:
-                pass
-            return
-
-        context = TranscriptionRunContext(
-            mode="free",
-            session_id=getattr(self.window, "current_session_id", None),
-            script_name=getattr(self.window, "current_script_name", "Free Speak"),
-            script_text=getattr(self.window, "current_script_text", ""),
-            audio_path=getattr(self.window, "current_audio_path", None),
-        )
-        self._active_free_context = context
-        self._mark_run_started(context)
-
-        self.window.free_worker = FreeTranscribeWorker(
-            self.model,
-            audio_input,
-            self.window,
-            options=self.get_whisper_options(free_speak=True),
-        )
-        context.worker = self.window.free_worker
-        self._received_segments_this_run = False
-
-        try:
-            self.window.free_worker.partial.connect(
-                lambda hyp, ctx=context: self.on_free_transcription_partial(hyp, ctx)
-            )
-            self.window.free_worker.completed_with_segments.connect(
-                lambda hyp, segments, ctx=context: (
-                    self.on_free_transcription_done_with_segments(hyp, segments, ctx)
-                )
-            )
-        except Exception:
-            pass
-        self.window.free_worker.completed.connect(
-            lambda hyp, ctx=context: self.on_free_transcription_done(hyp, ctx)
-        )
-        self.window.free_worker.failed.connect(
-            lambda msg, ctx=context: self.on_free_worker_failed(msg, ctx)
-        )
-        self.window.free_worker.finished.connect(self._finish_worker)
-        self.window.free_worker.start()
-
-    # ------------------------ slots: scoring -----------------------
-
-    def on_transcription_partial(
-        self, hyp: str, context: Optional[TranscriptionRunContext] = None
-    ) -> None:
-        """
-        Show accumulated chunk output while a long script-based transcription runs.
-        Final scoring/highlighting happens in the completed slots.
-        """
-        if context is None:
-            context = self._active_score_context
-        if not self._is_context_visible(context):
-            return
-        try:
-            self.window.transcript_txt.setPlainText(str(hyp))
-            self._clear_transcript_sync()
-            self._sync_export_button()
-            self.window.metrics_label.setText("Scoring... transcribed partial audio")
-        except Exception:
-            pass
-
-    def on_transcription_done(
-        self,
-        hyp: str,
-        err: float,
-        clar: float,
-        score: float,
-        context: Optional[TranscriptionRunContext] = None,
-    ) -> None:
-        """
-        Completion without segments. Show WER/CER/Clarity/Score,
-        compute error spans (no timing metrics available).
-        """
-        t_ui0 = time.perf_counter()
-        if context is None:
-            context = self._active_score_context
-        script_text = (
-            context.script_text if context is not None else self.window.current_script_text
-        )
-        script_name = (
-            context.script_name if context is not None else self.window.current_script_name
-        )
-        audio_path = (
-            context.audio_path if context is not None else self.window.current_audio_path
-        )
-        visible = self._is_context_visible(context)
-        received_segments = (
-            context.received_segments
-            if context is not None
-            else self._received_segments_this_run
-        )
-        cer_val = self._compute_cer(script_text, hyp)
-
-        if not received_segments:
-            if visible:
-                self.window.transcript_txt.setText(hyp)
-                self._clear_transcript_sync()
-                self._sync_export_button()
-
-                # Error highlights (no timestamps needed)
-                s_spans, t_spans = self.compute_error_spans(
-                    script_text,
-                    self.window.transcript_txt.toPlainText(),
-                )
-                self.window.set_error_highlights(s_spans, t_spans)
-                self.window.set_mistake_pairs(
-                    extract_mistake_pairs_for_display(
-                        script_text,
-                        self.window.transcript_txt.toPlainText(),
-                    )
-                )
-
-            # Persist if session exists
-            if context is not None and context.session_id is not None:
-                sess = db.update_session_scores(
-                    self.window.db,
-                    context.session_id,
-                    hyp,
-                    err,
-                    clar,
-                    score,
-                    segments_json=None,
-                    cer=cer_val,
-                )
-            else:
-                # Create session if none (rare path)
-                sess = db.add_session(
-                    self.window.db,
-                    script_name,
-                    script_text,
-                    audio_path or "",
-                    hyp,
-                    err,
-                    clar,
-                    score,
-                    segments_json=None,
-                    cer=cer_val,
-                )
-                if context is not None:
-                    context.session_id = sess.id
-                if visible:
-                    self.window.current_session_id = sess.id
-                self._add_history_item(sess)
-            try:
-                if sess is not None:
-                    events = extract_error_events(script_text, hyp)
-                    db.replace_session_errors(
-                        self.window.db,
-                        sess.id,
-                        sess.timestamp,
-                        sess.script_name,
-                        events,
-                    )
-            except Exception:
-                pass
-            self._mark_score_run_finished(context, sess)
-
-            timing = self._get_worker_timing(
-                free_speak=False,
-                worker=context.worker if context is not None else None,
-            )
-            timing_text = self._format_timing_text(timing, time.perf_counter() - t_ui0)
-            if visible:
-                self.window.metrics_label.setText(
-                    f"Score: {score:.2f}/5 | WER: {err:.2%} | "
-                    f"CER: {cer_val:.2%} | Clarity: {clar:.2%}"
-                )
-                self._set_timing_text(timing_text)
-
-        self._finish_worker()
-
-    def on_transcription_done_with_segments(
-        self,
-        hyp: str,
-        err: float,
-        clar: float,
-        score: float,
-        segments: object,
-        context: Optional[TranscriptionRunContext] = None,
-    ) -> None:
-        """
-        Completion with segments: compute extended metrics + spans,
-        update UI, and persist.
-        """
-        t_ui0 = time.perf_counter()
-        if context is None:
-            context = self._active_score_context
-        script_text = (
-            context.script_text if context is not None else self.window.current_script_text
-        )
-        script_name = (
-            context.script_name if context is not None else self.window.current_script_name
-        )
-        audio_path = (
-            context.audio_path if context is not None else self.window.current_audio_path
-        )
-        visible = self._is_context_visible(context)
-        try:
-            self._received_segments_this_run = True
-            if context is not None:
-                context.received_segments = True
-
-            seg_list = list(segments) if isinstance(segments, list) else []
-            # Augment and compute extended metrics
-            segs_aug, artic_rate, pause_ratio, filled_cnt, avg_conf = (
-                self._augment_segments_and_fluency(seg_list, hyp)
-            )
-            cer_val = self._compute_cer(script_text, hyp)
-
-            # Build transcript display and time ranges
-            txt, segs_built, ranges, active_idx = build_transcript_from_segments(
-                segs_aug
-            )
-
-            if visible:
-                self.window.transcript_segments = segs_aug
-                self.window.transcript_segment_ranges = ranges
-                self.window.transcript_active_index = active_idx
-                self.window.transcript_txt.setPlainText(txt)
-                self._sync_export_button()
-
-                # Error highlights (script vs built transcript text)
-                s_spans, t_spans = self.compute_error_spans(script_text, txt)
-                self.window.set_error_highlights(s_spans, t_spans)
-                self.window.set_mistake_pairs(
-                    extract_mistake_pairs_for_display(
-                        script_text,
-                        txt,
-                    )
-                )
-
-                # Label with extended metrics
-                avg_conf_txt = f"{avg_conf:.0%}" if avg_conf is not None else "-"
-                self.window.metrics_label.setText(
-                    "Score: "
-                    f"{score:.2f}/5 | WER: {err:.2%} | CER: {cer_val:.2%} | "
-                    f"Clarity: {clar:.2%} | Rate: {artic_rate:.0f} wpm | "
-                    f"Pauses: {pause_ratio:.0%} | Conf: {avg_conf_txt}"
-                )
-
-                # Initial playhead highlight along with base selections
-                self.window.transcript_active_index = -1
-                self.window.transcript_active_index = highlight_transcript_at_time(
-                    self.window.transcript_txt,
-                    self.window.transcript_segment_ranges,
-                    0.0,
-                    self.window.transcript_active_index,
-                    base_selections=self.window.transcript_error_selections,
-                )
-
-            # Persist
-            segments_json = json.dumps(segs_aug)
-            if context is not None and context.session_id is not None:
-                sess = db.update_session_scores(
-                    self.window.db,
-                    context.session_id,
-                    hyp,
-                    err,
-                    clar,
-                    score,
-                    segments_json=segments_json,
-                    cer=cer_val,
-                    artic_rate=artic_rate,
-                    pause_ratio=pause_ratio,
-                    filled_pauses=filled_cnt,
-                    avg_conf=avg_conf,
-                )
-            else:
-                sess = db.add_session(
-                    self.window.db,
-                    script_name,
-                    script_text,
-                    audio_path or "",
-                    hyp,
-                    err,
-                    clar,
-                    score,
-                    segments_json=segments_json,
-                    cer=cer_val,
-                    artic_rate=artic_rate,
-                    pause_ratio=pause_ratio,
-                    filled_pauses=filled_cnt,
-                    avg_conf=avg_conf,
-                )
-                if context is not None:
-                    context.session_id = sess.id
-                if visible:
-                    self.window.current_session_id = sess.id
-                self._add_history_item(sess)
-            try:
-                if sess is not None:
-                    events = extract_error_events(script_text, hyp)
-                    db.replace_session_errors(
-                        self.window.db,
-                        sess.id,
-                        sess.timestamp,
-                        sess.script_name,
-                        events,
-                    )
-            except Exception:
-                pass
-            self._mark_score_run_finished(context, sess)
-
-            timing = self._get_worker_timing(
-                free_speak=False,
-                worker=context.worker if context is not None else None,
-            )
-            timing_text = self._format_timing_text(timing, time.perf_counter() - t_ui0)
-            avg_conf_txt = f"{avg_conf:.0%}" if avg_conf is not None else "-"
-            if visible:
-                self.window.metrics_label.setText(
-                    "Score: "
-                    f"{score:.2f}/5 | WER: {err:.2%} | CER: {cer_val:.2%} | "
-                    f"Clarity: {clar:.2%} | Rate: {artic_rate:.0f} wpm | "
-                    f"Pauses: {pause_ratio:.0%} | Conf: {avg_conf_txt}"
-                )
-                self._set_timing_text(timing_text)
-
-        except Exception:
-            # Fallback: show plain hypothesis if anything goes wrong
-            if visible:
-                self.window.transcript_txt.setText(hyp)
-                self._clear_transcript_sync()
-                self._sync_export_button()
-        finally:
-            self._finish_worker()
-
-    # ------------------------ slots: free speak ---------------------
-
-    def on_free_transcription_partial(
-        self, hyp: str, context: Optional[TranscriptionRunContext] = None
-    ) -> None:
-        """
-        Show accumulated chunk output while a long Free Speak transcription runs.
-        Final fluency/confidence metrics happen in the completed slots.
-        """
-        if context is None:
-            context = self._active_free_context
-        if not self._is_context_visible(context):
-            self.window.last_transcript_text = str(hyp)
-            return
-        try:
-            self.window.last_transcript_text = str(hyp)
-            self.window.transcript_txt.setPlainText(str(hyp))
-            self._clear_transcript_sync()
-            self._sync_export_button()
-            self.window.set_mistake_pairs([])
-            self.window.metrics_label.setText(
-                "Transcribing... (Free Speak) transcribed partial audio"
-            )
-        except Exception:
-            pass
-
-    def on_free_transcription_done(
-        self, hyp: str, context: Optional[TranscriptionRunContext] = None
-    ) -> None:
-        """
-        Free Speak without segments. No WER/CER. Keep transcript visible.
-        """
-        t_ui0 = time.perf_counter()
-        if context is None:
-            context = self._active_free_context
-        visible = self._is_context_visible(context)
-        received_segments = (
-            context.received_segments
-            if context is not None
-            else self._received_segments_this_run
-        )
-        self.window.last_transcript_text = hyp
-
-        if not received_segments and visible:
-            self.window.transcript_txt.setText(hyp)
-            self._clear_transcript_sync()
-            self._sync_export_button()
-            self.window.set_mistake_pairs([])
-            timing = self._get_worker_timing(
-                free_speak=True,
-                worker=context.worker if context is not None else None,
-            )
-            timing_text = self._format_timing_text(timing, time.perf_counter() - t_ui0)
-            self.window.metrics_label.setText("Transcript ready (Free Speak)")
-            self._set_timing_text(timing_text)
-        if visible:
-            self.window.btn_save.setEnabled(True)
-        self._mark_free_run_finished()
-        self._finish_worker()
-
-    def on_free_transcription_done_with_segments(
-        self,
-        hyp: str,
-        segments: object,
-        context: Optional[TranscriptionRunContext] = None,
-    ) -> None:
-        """
-        Free Speak with segments: compute fluency/confidence only.
-        """
-        t_ui0 = time.perf_counter()
-        if context is None:
-            context = self._active_free_context
-        visible = self._is_context_visible(context)
-        try:
-            self._received_segments_this_run = True
-            if context is not None:
-                context.received_segments = True
-            self.window.last_transcript_text = hyp
-
-            seg_list = list(segments) if isinstance(segments, list) else []
-            segs_aug, artic_rate, pause_ratio, filled_cnt, avg_conf = (
-                self._augment_segments_and_fluency(seg_list, hyp)
-            )
-
-            txt, segs_built, ranges, active_idx = build_transcript_from_segments(
-                segs_aug
-            )
-            if visible:
-                self.window.transcript_segments = segs_aug
-                self.window.transcript_segment_ranges = ranges
-                self.window.transcript_active_index = active_idx
-                self.window.transcript_txt.setPlainText(txt)
-                self._sync_export_button()
-                self.window.set_mistake_pairs([])
-
-            avg_conf_txt = f"{avg_conf:.0%}" if avg_conf is not None else "-"
-            if visible:
-                self.window.metrics_label.setText(
-                    "Transcript ready (Free Speak) | "
-                    f"Rate: {artic_rate:.0f} wpm | "
-                    f"Pauses: {pause_ratio:.0%} | Conf: {avg_conf_txt}"
-                )
-            timing = self._get_worker_timing(
-                free_speak=True,
-                worker=context.worker if context is not None else None,
-            )
-            timing_text = self._format_timing_text(timing, time.perf_counter() - t_ui0)
-            if visible:
-                self._set_timing_text(timing_text)
-
-            if visible:
-                self.window.transcript_active_index = -1
-                self.window.transcript_active_index = highlight_transcript_at_time(
-                    self.window.transcript_txt,
-                    self.window.transcript_segment_ranges,
-                    0.0,
-                    self.window.transcript_active_index,
-                    base_selections=self.window.transcript_error_selections,
-                )
-            # Persistence happens on Save in free speak mode
-        except Exception:
-            if visible:
-                self.window.transcript_txt.setText(hyp)
-                self._clear_transcript_sync()
-                self._sync_export_button()
-                self.window.set_mistake_pairs([])
-        finally:
-            if visible:
-                self.window.btn_save.setEnabled(True)
-            self._mark_free_run_finished()
-            self._finish_worker()
-
-    # ------------------------ error handling -----------------------
-
-    def on_worker_failed(
-        self, msg: str, context: Optional[TranscriptionRunContext] = None
-    ) -> None:
-        if context is None:
-            context = self._active_score_context
-        if self._is_context_visible(context):
-            self.window.metrics_label.setText(f"Scoring failed: {msg}")
-            self._set_timing_text("Timing: failed")
-        self._mark_score_run_finished(context, None, failed=True)
-        self._finish_worker()
-
-    def on_free_worker_failed(
-        self, msg: str, context: Optional[TranscriptionRunContext] = None
-    ) -> None:
-        if context is None:
-            context = self._active_free_context
-        if self._is_context_visible(context):
-            self.window.metrics_label.setText(f"Transcription failed: {msg}")
-            self._set_timing_text("Timing: failed")
-            try:
-                self.window.btn_save.setEnabled(False)
-            except Exception:
-                pass
-        self._mark_free_run_finished(failed=True)
-        self._finish_worker()
-
-    # ------------------------ utilities ----------------------------
-
-    def _clear_transcript_sync(self) -> None:
-        """
-        Clear transcript sync but keep base error selections (if any) visible.
-        """
-        self.window.transcript_segments = None
-        self.window.transcript_segment_ranges = []
-        self.window.transcript_active_index = -1
-        try:
-            self.window.transcript_txt.setExtraSelections(
-                list(self.window.transcript_error_selections or [])
-            )
-        except Exception:
-            pass
